Alpha 0.2
This is an alpha 0.2 version of the game.

There is a lot of errors to correct and code to be re-written.

+new:Enemies now spawn in random sizes based on the size of the player.

+new:A growth factor was implemented to slow down the growth as it was too fast.

+new:The camera position changes as you grow so you can see a wider field of view.

*fixed: enemies were unable to spawn after initial spawns

+++ Game/Stellar Growth/Assets/Scripts/Mass.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Mass : MonoBehaviour
 {
 
-    public int objectMass = 1;
+    public float objectMass = 1;
     private Vector3 size;
 
     // Start is called before the first frame update
@@ -22,12 +22,12 @@ public class Mass : MonoBehaviour
     }
 
 
-    public void setMass(int mass)
+    public void setMass(float mass)
     {
         this.objectMass = mass;
     }
 
-    public int getMass()
+    public float getMass()
     {
         return this.objectMass;
     }
